pyproject.toml
[project]
name = "lexrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "jq>=1.10.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.34",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.4",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.1",
    "tiktoken>=0.11.0",
    "unstructured[pdf]>=0.18.15",
]
